Rewrite NorwegianDateUtil to use java.time
java.util.Date and java.util.Calendar is old now and we don't want to
use that any more.

Breaking changes:

We use ZonedDateTime in stead of Date to represent a date to check
for working days.

The list of holidays are now a NavigableSet<LocalDate> and not a Date[].

Co-authored-by: tobiasgwaaler
Co-authored-by: runeflobakk
Co-authored-by: martin-jackson
Co-authored-by: hhaga
Co-authored-by: fredrbus